Move Loadable into rx helper package
