+++ gnd/src/main/java/com/google/android/gnd/MainViewModel.java
@@ -22,8 +22,8 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import com.google.android.gnd.model.Project;
 import com.google.android.gnd.repository.FeatureRepository;
-import com.google.android.gnd.repository.Loadable;
 import com.google.android.gnd.repository.ProjectRepository;
+import com.google.android.gnd.rx.Loadable;
 import com.google.android.gnd.ui.common.AbstractViewModel;
 import com.google.android.gnd.ui.common.Navigator;
 import com.google.android.gnd.ui.common.SharedViewModel;
