+++ gnd/src/main/java/com/google/android/gnd/ui/observationdetails/ObservationDetailsViewModel.java
@@ -22,8 +22,8 @@ import androidx.lifecycle.LiveDataReactiveStreams;
 import com.google.android.gnd.model.feature.Feature;
 import com.google.android.gnd.model.form.Form;
 import com.google.android.gnd.model.observation.Observation;
-import com.google.android.gnd.repository.Loadable;
 import com.google.android.gnd.repository.ObservationRepository;
+import com.google.android.gnd.rx.Loadable;
 import com.google.android.gnd.ui.common.AbstractViewModel;
 import io.reactivex.Flowable;
 import io.reactivex.processors.BehaviorProcessor;
